Cleanup Factory Registry and related classes
